Fixed minor compiler warnings.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1325 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: f551cfef238702136f7955b76a1e5dd6e844c303

+++ src/core/Peridigm.cpp
@@ -1100,7 +1100,7 @@ void PeridigmNS::Peridigm::jacobianDiagnostics(Teuchos::RCP<NOX::Epetra::Group> 
 
   if(!noxGroup->isConditionNumber()){
     int conditionNumberMaxIters = 10000;
-    int conditionNumberTolerance = 1.0e3;
+    double conditionNumberTolerance = 1.0e3;
     int conditionNumberKrylovSubspaceSize = 100;
     int conditionNumberPrintOutput = false;
     returnValue = noxGroup->computeJacobianConditionNumber(conditionNumberMaxIters,
